This fixes several minor warnings uncovered by the CentOS 5.5 version of the C++ compiler.  It also adds Derek's ${MAKE} fix.
git-svn-id: https://tricklab.jsc.nasa.gov/svn/repos/dcapp/trunk@38 caffc767-fac6-498e-a973-f5283f39b8c3

+++ dyn_elements.cc
@@ -45,7 +45,7 @@ int update_dyn_elements(struct node *list)
                         {
                             current->object.pixelstream.buffercount = current->object.pixelstream.shm->buffercount;
                             rewind(current->object.pixelstream.fp);
-                            newh = current->object.pixelstream.shm->width * (*(current->info.h)) / (*(current->info.w));
+                            newh = (size_t)((float)(current->object.pixelstream.shm->width) * (*(current->info.h)) / (*(current->info.w)));
                             if (newh > current->object.pixelstream.shm->height)
                             {
                                 origbytes = current->object.pixelstream.shm->width * current->object.pixelstream.shm->height * 3;
